Agregando sweetalert al proyecto

+++ routes/web.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\OrderController;
 use App\Http\Controllers\ProductController;
 use App\Http\Controllers\ProviderController;
 use App\Http\Controllers\SaleController;
+use App\Http\Controllers\UserController;
 use App\Models\Product;
 use App\Models\Provider;
 use App\Models\Sale;
@@ -30,38 +31,33 @@ Route::get('/', function () {
     return view('index');
 })->middleware('auth');
 
-// Usuarios
-// Route::get('/login', [LoginController::class, 'login'])->name('login');
-// Route::post('/auth', [LoginController::class, 'authenticate'])->name('auth');
+// Autenticacion
+Route::get('/login', [LoginController::class, 'login'])->name('login');
+Route::post('/auth', [LoginController::class, 'authenticate'])->name('auth');
+Route::get('/logout', [LoginController::class, 'logout'])->name('logout');
 
 // Productos
 // Las rutas agregadas se colocan antes del resource para evitar conflictos.
-Route::get('products/listDelete', [ProductController::class, 'listDelete'])->name('products.listDelete');
-Route::get('products/restoreProduct/{product}', [ProductController::class, 'restoreProduct'])->name('products.restoreProduct');
+Route::get('products/listDelete', [ProductController::class, 'listDelete'])->name('products.listDelete')->middleware('auth');
+Route::get('products/restoreProduct/{product}', [ProductController::class, 'restoreProduct'])->name('products.restoreProduct')->middleware('auth');
 
-Route::resource('products', ProductController::class);
+Route::resource('products', ProductController::class)->middleware('auth');
 
 // Proveedores
-Route::resource('providers', ProviderController::class);
+Route::resource('providers', ProviderController::class)->middleware('auth');
 
 // Ventas
-Route::resource('sales', SaleController::class);
+Route::resource('sales', SaleController::class)->middleware('auth');
 
 // Pedidos
-Route::delete('orders/destroyItem/{item}', [OrderController::class, 'destroyItem'])->name('orders.destroyItem');
-Route::get('orders/cambiarEstado/{item}', [OrderController::class, 'cambiarEstado'])->name('orders.cambiarEstado');
-Route::get('orders/cargarItems/{item}', [OrderController::class, 'cargarItems'])->name('orders.cargarItems');
-Route::resource('orders', OrderController::class);
+Route::delete('orders/destroyItem/{item}', [OrderController::class, 'destroyItem'])->name('orders.destroyItem')->middleware('auth');
+Route::get('orders/cambiarEstado/{item}', [OrderController::class, 'cambiarEstado'])->name('orders.cambiarEstado')->middleware('auth');
+Route::get('orders/cargarItems/{item}', [OrderController::class, 'cargarItems'])->name('orders.cargarItems')->middleware('auth');
+Route::resource('orders', OrderController::class)->middleware('auth');
 
 // Categorias
-Route::resource('categories', CategoryController::class);
-
+Route::resource('categories', CategoryController::class)->middleware('auth');
 
 // Usuarios
-Route::get('/usuarios/nuevo-usuario', function () {
-    return view('usuarios.nuevo-usuario');
-});
+Route::resource('users', UserController::class)->middleware('auth');
 
-Route::get('/usuarios/listado-usuarios', function () {
-    return view('usuarios.listado-usuarios');
-});
